Support generic attributes in WALKey
Adds a facility for decorating WALKeys with generic attributes, for when a
coprocessor application wants to decorate a WALedit with extension data for
use in a local WAL observer, and/or a custom replication endpoint at either
source or sink.

Adds a RegionObserver hook suitable for decorating WALKeys or WALEdits
about to be appended to the WAL, called out of HRegion operations.