HBASE-28919 Soft drop for destructive table actions
Although HFiles are copied to the archive in a destructive schema change, recovery
scenarios are not automatic and involve some operator labor to reconstruct the
table and re-import the archived data. We can easily prevent the deletion of the
HFiles of a deleted table or column family by taking a snapshot of the table
immediately prior to any destructive schema actions. We also set a TTL on the
snapshot so housekeeping of unwanted HFiles remains no touch. Because we take a
table snapshot all table structure and metadata is also captured and saved so fast
recovery is possible, as either a restore from snapshot, or a clone from snapshot
to a new table.

Existing site configuration property prerequisites:

* hbase.snapshot.enabled = true ( default is true )

New site configuration properties:

* hbase.snapshot.before.destructive.action.enabled = true ( default is false )
* hbase.snapshot.before.destructive.action.ttl = <integer>, in seconds ( default 86400 (one day) )